#152 #147 ResolutionStrategy is now an interface
